fix: redis cache and tests files

+++ src/models/Usuario.ts
@@ -5,38 +5,49 @@ import { env } from '../env'
 const prisma = new PrismaClient()
 
 export const getUsers = async () => {
-	// Obtém os dados do cache com await
-	const cache = await client.get(env.REDIS_CACHED_KEY)
-	if (cache) {
-		console.log('cache')
-		const cachedUsers = JSON.parse(cache, (key, value) => {
-			// Converte strings de data de volta para objetos Date
-			if (key === 'createdAt' || key === 'updatedAt') {
-				return new Date(value)
-			}
-			return value
-		}) as Usuario[]
-		return cachedUsers
+	if (client.isOpen) {
+		// Obtém os dados do cache com await
+		const cache = await client.get(env.REDIS_CACHED_KEY)
+		if (cache) {
+			console.log('cache')
+			const cachedUsers = JSON.parse(cache, (key, value) => {
+				// Converte strings de data de volta para objetos Date
+				if (key === 'createdAt' || key === 'updatedAt') {
+					return new Date(value)
+				}
+				return value
+			}) as Usuario[]
+			return cachedUsers
+		}
 	}
 	// Obtém os dados do banco de dados
 	const users = await prisma.usuario.findMany()
 
-	// Salva os dados no cache por 60 segundos
-	const cachedUsers = JSON.stringify(users)
-	await client.set(env.REDIS_CACHED_KEY, cachedUsers).catch(err => console.error('Erro ao salvar no cache:', err))
+	if (client.isOpen) {
+		// Salva os dados no cache
+		const cachedUsers = JSON.stringify(users)
+		await client.set(env.REDIS_CACHED_KEY, cachedUsers).catch(err => console.error('Erro ao salvar no cache:', err))
+	}
+
 	return users
 }
 
 export const createUser = async (data: Pick<Usuario, 'nome' | 'email'>) => {
 	const result = await prisma.usuario.create({ data })
-	// Remove o cache caso o usuário seja criado
-	await client.del(env.REDIS_CACHED_KEY)
+	if (client.isOpen) {
+		// Remove o cache caso o usuário seja criado
+		await client.del(env.REDIS_CACHED_KEY)
+	}
+
 	return result
 }
 
 export const updateUser = async (id: string, data: Pick<Usuario, 'nome' | 'email'>) => {
-	// Remove o cache caso o usuário seja atualizado
-	await client.del(env.REDIS_CACHED_KEY)
+	if (client.isOpen) {
+		// Remove o cache caso o usuário seja atualizado
+		await client.del(env.REDIS_CACHED_KEY)
+	}
+
 	return await prisma.usuario.update({
 		data,
 		where: { id },
@@ -44,8 +55,10 @@ export const updateUser = async (id: string, data: Pick<Usuario, 'nome' | 'email
 }
 
 export const deleteUser = async (id: string) => {
-	// Remove o cache caso o usuário seja deletado
-	await client.del(env.REDIS_CACHED_KEY)
+	if (client.isOpen) {
+		// Remove o cache caso o usuário seja deletado
+		await client.del(env.REDIS_CACHED_KEY)
+	}
 	return await prisma.usuario.delete({
 		where: { id: id },
 	})
